Add files via upload
Programmed animations onto the player and missile. Added missile launchers but they are not working 100% as intended as there is no delay between the first missile disappearing and the 2nd missile appearing. Missing a jump animation for the player and the missiles do not face the correct way at the moment when moving towards the player but that can be fixed later.